requirements.txt
streamlit>=1.37
pandas>=2.2
openpyxl>=3.1
python-dotenv>=1.0
pydantic>=2.7
email-validator>=2.1
phonenumbers>=8.13
faker>=26.0
schedule>=1.2
typer>=0.12
langchain>=0.2
langgraph>=0.2
